# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KIPS School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [active, inactive, left, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [paid, unpaid, partial, null]
        - name: month
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
    post:
      operationId: createFee
      tags: [fees]
      summary: Create fee record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Fee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /fees/{id}/pay:
    post:
      operationId: payFee
      tags: [fees]
      summary: Mark fee as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeePayment"
      responses:
        "200":
          description: Fee paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
    post:
      operationId: createCertificate
      tags: [certificates]
      summary: Generate certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Certificate generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, name]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, teacher, student]
        name:
          type: string
        email:
          type: ["string", "null"]
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    Student:
      type: object
      required: [id, admissionNumber, name, classId, status, createdAt]
      properties:
        id:
          type: integer
        admissionNumber:
          type: string
        name:
          type: string
        fatherName:
          type: ["string", "null"]
        motherName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        classId:
          type: integer
        className:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        rollNumber:
          type: ["string", "null"]
        feeAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, inactive, left]
        imageUrl:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, classId]
      properties:
        name:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female]
        address:
          type: string
        phone:
          type: string
        emergencyContact:
          type: string
        classId:
          type: integer
        section:
          type: string
        rollNumber:
          type: string
        feeAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, left]
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        fatherName:
          type: string
        motherName:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        address:
          type: string
        phone:
          type: string
        emergencyContact:
          type: string
        classId:
          type: integer
        section:
          type: string
        rollNumber:
          type: string
        feeAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, left]
...
    FeeRecord:
      type: object
      required: [id, studentId, amount, month, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        admissionNumber:
          type: ["string", "null"]
        className:
          type: ["string", "null"]
        amount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        month:
          type: string
        dueDate:
          type: string
        paidDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [paid, unpaid, partial]
        fine:
          type: number
        createdAt:
          type: string
    FeeInput:
      type: object
      required: [studentId, amount, month, dueDate]
      properties:
        studentId:
          type: integer
        amount:
          type: number
        month:
          type: string
        dueDate:
          type: string
        fine:
          type: number
    FeePayment:
      type: object
      required: [paidAmount]
      properties:
        paidAmount:
          type: number
...
    Certificate:
      type: object
      required: [id, studentId, type, issuedDate, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        type:
          type: string
          enum: [leaving, birth, bonafide, character, result, fee_clearance]
        issuedDate:
          type: string
        certificateNumber:
          type: string
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    CertificateInput:
      type: object
      required: [studentId, type, issuedDate]
      properties:
        studentId:
          type: integer
        type:
          type: string
          enum: [leaving, birth, bonafide, character, result, fee_clearance]
        issuedDate:
          type: string
        remarks:
          type: string